Adds global key event monitoring
Introduces a global key event monitor using the `rdev` crate to capture key presses and releases.

Adds `serialize` feature to `rdev` dependency to resolve build issues.

The monitor can be toggled by clicking a dedicated area in the UI.

Pressing `Ctrl+Alt+0` stops the key monitor.
On macOS, stops the run loop; on other platforms, panics with a custom `MonitorStop` error.

This is implemented in a separate thread to avoid blocking the main UI thread.