Improves client-server communication & UI
Adds functionality for the client to send mouse movements to the server via QUIC.

The client sends mouse events to server using rmp-serde for serialization over uni-directional streams.
Introduces a shared library for the MouseMove data structure.
Adds the ability for user to press enter while focused on the IP or PORT entry to automatically click the 'connect' button.
Adds a keybinding to stop the key monitor.

Closes connections gracefully upon completion or error.