Allows stream finishing after Ctrl+Alt+0
Modifies the key monitor to allow finishing of mouse and keyboard streams after detecting the Ctrl+Alt+0 key combination.

This change prevents a crash caused by attempting to use a finished stream by wrapping `mouse_stream` and `keyboard_stream` in an `Option`.
This allows the streams to be consumed using `.take()` and dropped.